Document-Summarizer-main/requirements.txt
fastapi
google.generativeai
Pillow
